Replace ViewModel with launchUiStateProducer in stateful screen composables.
This backs out ViewModel until scoping can be resolved and replace with an effect similar to the previous architecture.

Re-add ViewModel when https://issuetracker.google.com/issues/165642391 is fixed

This commit adds a @Composeable function for loading data from a data source that exposes suspend funs [launchUiStateProducer].

Since the previous ViewModel refactor extracted a stateless and stateful composable for each screen, we are able to completely
change how state management is implemented without modifying the UI code for the screen. On [HomeScreen] [ArticleScreen] and
[InterestsScreen] the refactor could be limited to edits in a single composable.

launchUiStateProducer launches a coroutine that fetches data and exposes the ability to refresh and clear error conditions as
events to ui code:

```
val (uiState, onRefresh, onClearError) = launchUiStateProducer(dataSource, resourceId) { loadData(resourceId) }
```

It is uses [launchInComposition] to safely manage coroutines inside of composition.

Changes:
- Created launchUiStateProducer effect to safely call suspending data producers from compose and return refreshable UI state
- Removed ViewModel from HomeScreen, InterestsScreen and ArticleScreen
- Refactored stateful screen composables to use the new launchUiStateProducer to load data
- Add mutex around in-memory read/edit/write operations in repositoryies.