[Owl] Fix tests never ending
The TestImageLoader was set to return the same
drawable instance for every request. This causes issues
though as Drawables are stateful, and therefore can't
be used in multiple places at once.

Fixed by returning a new instance for each request.
Also remembered the image loader for a small perf
improvement.

Change-Id: Ide1996527fb3f8c2918c391532454aa72c7b401a