[Crane] Set the onClick label for DayContainer
This specifies the onClick label to use for the individual calendar
days. I also stopped passing the selected boolean to Surface. When it is
passed, that uses the selectable version of Surface which announces the
composable as a Tab and automatically states "selected" at the start of
the announcement when selected==true but does not announce anything at
the start when not selected.

Instead, it's now using the clickable version of Surface, so our
announcements are like:

"{Selected/not selected}, January 1st 2020, Button, Double-tap to
select"

This seems reasonable, since the other Roles are Checkbox, Switch,
RadioButton, and Image, all of which seem less correct than Button.