Fixed toolbar and status bar
Jetsurvey has had an issue since before this migration where the status
bar text/icons were too light in light mode. This fixes that my using
Accompanist, but I didn't enable edge-to-edge yet. I also missed that m3
has a custom TopAppBar to use when the text is center-aligned, so I
switched to that one.