Fix for status bar icon colors
JetLagged only supports a light theme, which means the status bar icons
should always use a light theme (meaning they have to be dark). I added
the statusBarStyle to the enableEdgeToEdge call to force a light theme.
The bottom nav also needs to be fixed but that's more involved because
every screen needs to have enough padding to push the content into view
when scrolled.