Added window insets to the the main screen
Since the headder and the body (FlowRow) are sibling elements, we have
to explicitly exclude the top inset, which we get by combining the
bottom and the sides (horizontal). Tested with all the developer options
notch settings and with both gesture nav and 3 button nav.

I did not override the navigationBarStyle since the default actually
seems to work well. With three button nav, it has a partially
transparent light background that helps the buttons stand out when
content is behind them. With gesture nav, the bar stands out enough on
its own.

This fixes #1450.